+++ frontend/src/EDR/components/Table.tsx
@@ -12,6 +12,7 @@ import { DetailedTrain } from "../functions/trainDetails";
 import {format} from "date-fns";
 import {TrainTimetableModal} from "./TrainTimetableModal";
 import i18n from "../../i18n";
+import { getTimetableStartingFromHour } from "../functions/trainsTable";
 
 export type Bounds = {
     firstColBounds: RectReadOnly;
@@ -70,33 +71,11 @@ export const EDRTable: React.FC<Props> = ({
     if (!trainsWithDetails || !post) return null;
     const postCfg = postConfig[post];
     const showStopColumn = timetable.length > 0 && timetable.some((row: any) => row.platform || Math.ceil(parseInt(row.layover)) !== 0);
-
-    const getTimetableStartingFromHour = (timetable: TimeTableRow[], startHour: string) => {      
-        // Find the index of the first time in the timetable that is equal to or
-        // greater than the start hour
-        const startIndex = timetable.findIndex((time: TimeTableRow) => time.scheduled_arrival >= startHour);
-        
-        if (startIndex === -1) {
-          // If there is no time in the timetable that is equal to or greater than the
-          // start hour, add the times from the beginning of the 24-hour period up to
-          // (but not including) the start hour, as well as all times from the start
-          // hour until the end of the 24-hour period
-            return [
-                ...timetable.slice(0, timetable.findIndex((time: TimeTableRow) => time.scheduled_arrival === startHour)),
-                ...timetable.slice(timetable.findIndex((time: TimeTableRow) => time.scheduled_arrival === startHour))
-            ];
-        } else {
-          // If there is a time in the timetable that is equal to or greater than the
-          // start hour, return a new array that starts from the start index and
-          // includes all times up to the end of the 24-hour period, as well as all
-          // times from the beginning of the 24-hour period up to (but not including)
-          // the start index
-            return [
-                ...timetable.slice(startIndex),
-                ...timetable.slice(0, startIndex)
-            ];
-        }
-    }
+    const timetableStartHour = formatTime(nowUTC(serverTzOffset, 1), i18n.language, {
+        hour: '2-digit',
+        minute: '2-digit',
+        hour12: false
+    });
 
     return <div>
         <SimRailMapModal serverCode={serverCode} trainId={mapModalTrainId} setModalTrainId={setMapModalTrainId} />
@@ -120,11 +99,7 @@ export const EDRTable: React.FC<Props> = ({
                 {timetable.length > 0
                     ? getTimetableStartingFromHour(
                         timetable,
-                        formatTime(nowUTC(serverTzOffset, 1), i18n.language, {
-                            hour: '2-digit',
-                            minute: '2-digit',
-                            hour12: false
-                        }))
+                        timetableStartHour)
                         .filter((tt) => filter ? 
                             // Remove spaces, trim not enough since humans usually use space after a separator
                             filter.replace(/\s+/g, '')
